feat(issue-270-test-ownership-contract): slice 2 — relocate blueprint-author tests, update required_files
