+++ scripts/bin/infra/contract_test_fast.sh
@@ -35,8 +35,8 @@ base_tests=(
   "tests/infra/test_runtime_identity_contract_cli.py"
   "tests/infra/test_argocd_repo_contract_cli.py"
   "tests/infra/test_state_artifact_contract.py"
-  "tests/infra/test_root_dir_resolution.py"
-  "tests/infra/test_optional_module_required_env_contract.py"
+  "tests/blueprint/test_root_dir_resolution.py"
+  "tests/blueprint/test_optional_module_required_env_contract.py"
   "tests/infra/test_tooling_contracts.py"
 )
 template_source_only_tests=(
