+++ scripts/templates/blueprint/bootstrap/blueprint/contract.yaml
@@ -511,12 +511,8 @@ spec:
       - tests/e2e/test_smoke_endpoints.py
       - tests/e2e/test_vertical_slice.py
       - tests/infra/test_argocd_repo_contract_cli.py
-      - tests/infra/test_async_message_contracts.py
       - tests/infra/test_core_runtime_bootstrap.py
-      - tests/infra/test_optional_module_required_env_contract.py
       - tests/infra/test_optional_modules.py
-      - tests/infra/test_python_helper_extractions.py
-      - tests/infra/test_root_dir_resolution.py
       - tests/infra/test_runtime_credentials_eso.py
       - tests/infra/test_runtime_identity_contract_cli.py
       - tests/infra/test_sdd_asset_checker.py
