+++ tests/blueprint/test_optional_module_required_env_contract.py
