+++ tests/blueprint/test_root_dir_resolution.py
